requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0

# Web scraping
selenium>=4.0.0
beautifulsoup4>=4.11.0
webdriver-manager>=3.8.0

# NLP and ML
nltk>=3.8
spacy>=3.4.0
sentence-transformers>=2.2.0
transformers>=4.20.0
torch>=1.12.0
scikit-learn>=1.1.0

# Vector search
faiss-cpu>=1.7.0  # Use faiss-gpu if you have CUDA

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.64.0

# UI (optional)
streamlit>=1.22.0
flask>=2.2.0

# Development and testing
pytest>=7.0.0
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

plotly